# gateway/docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.GetUserByIdRequest:
    properties:
      id:
        type: string
    type: object
  models.GetUserByIdResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.GetUserByUsernameRequest:
    properties:
      username:
        type: string
    type: object
  models.GetUserByUsernameResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.GetUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/models.UserForGetUsers'
        type: array
    type: object
...
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  models.UpdateMeRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.UserForGetUsers:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the Messenger service
  title: Messenger API
  version: "1.0"
paths:
...
  /user/id:
    get:
      consumes:
      - application/json
      description: Retrieve user data by user ID.
      parameters:
      - description: id пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GetUserByIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserByIdResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User By ID
      tags:
      - users
  /user/me:
    get:
      consumes:
      - application/json
      description: Retrieve the currently authenticated user's data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User Information
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the currently authenticated user's data.
      parameters:
      - description: User Update Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User Information
      tags:
      - users
  /user/name:
    get:
      consumes:
      - application/json
      description: получить пользователя по username
      parameters:
      - description: username пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GetUserByUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserByUsernameResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetUserByUsername
      tags:
      - users
  /user/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"